Refactored data parsing into a separate file
Katherine and I worked out the important sections of parsing data and
put it into a separate r function in a separate file.  My motivation
here is that it will make the code more reproducible and less redundancy
in parsing.  Feel free to edit it or reinclude anything I removed.